Minimalist version that compiles and runs on Linux
A concise description of this system would look like this:
	Linux ironman 3.2.0-61-virtual #93-Ubuntu SMP Fri May 2 21:54:33 UTC 2014 x86_64 x86_64 x86_64 GNU/Linux
	g++ (Ubuntu/Linaro 4.6.3-1ubuntu5) 4.6.3
	cmake version 2.8.7

All the examples that were written to work on Windows but don't work in
Linux are moved to their own category of examples. Regrettably, that
includes the test cases.

All of the rest worked OOTB though, so it's not all bad: if my Wolverine
and Daredevil servers can also compile this, there's a pretty good
chance that the rest can be ported back as needed.